Completed: Task 13.16.
Make a route for the application /api/authors that returns the number of blogs for each author and the total number of likes. Implement the operation directly at the database level. You will most likely need the group by functionality, and the sequelize.fn aggregator function.

Bonus task: order the data returned based on the number of likes, do the ordering in the database query.