Completed: Exercise 13.20.
Now add functionality to the application to support the reading list.

Adding a blog to the reading list is done by making an HTTP POST to the path /api/readinglists, the request will be accompanied with the blog and user id.

Also modify the individual user route GET /api/users/:id to return not only the user's other information but also the reading list.

At this point, information about whether the blog is read or not does not need to be available.