Completed: Exercise 13.22.
Implement functionality in the application to mark a blog in the reading list as read. Marking as read is done by making a request to the PUT /api/readinglists/:id path.

The user can only mark the blogs in their own reading list as read. The user is identified as usual from the token accompanying the request.